DBZ-7260 Compeltely removed ReplicaSetDiscovery
